Store the http request duration as a float.

+++ app/decorators/check_time.rb
@@ -1,6 +1,6 @@
 class CheckTime < Decorator
 
   def call(response)
-    check.value.to_i >= response.duration
+    check.value.to_f >= response.duration
   end
 end
